% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpumat-interface.r
\name{as_cpumat}
\alias{as_cpumat}
\title{as_cpumat}
\usage{
as_cpumat(x, copy = TRUE)
}
\arguments{
\item{x}{R matrix.}

\item{copy}{Should the R data be copied? If \code{FALSE}, be careful!}
}
\value{
A cpumat object.
}
\description{
Convert an R matrix to a cpumat object.
}
