% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuvec.r
\name{as_cpuvec}
\alias{as_cpuvec}
\title{as_cpuvec}
\usage{
as_cpuvec(x, copy = TRUE)
}
\arguments{
\item{x}{R vector.}

\item{copy}{Should the R data be copied? If \code{FALSE}, be careful!}
}
\value{
A cpuvec object.
}
\description{
Convert an R vector to a cpumat object.
}
