% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpumat.r
\name{as_gpumat}
\alias{as_gpumat}
\title{as_gpumat}
\usage{
as_gpumat(card, x, copy = TRUE)
}
\arguments{
\item{card}{A GPU card object; the return of \code{card()}. See \code{?card}.}

\item{x}{R matrix.}

\item{copy}{Should the R data be copied? If \code{FALSE}, be careful!}
}
\value{
A gpumat object.
}
\description{
Convert an R matrix to a gpumat object.
}
