% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpimat.r
\name{as_mpimat}
\alias{as_mpimat}
\title{as_mpimat}
\usage{
as_mpimat(grid, x, copy = TRUE)
}
\arguments{
\item{x}{R matrix.}

\item{copy}{Should the R data be copied? If \code{FALSE}, be careful!}
}
\value{
A mpimat object.
}
\description{
Convert an R matrix to a mpimat object.
}
