% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{bar_tick}
\alias{bar_tick}
\title{Construct tick bars}
\usage{
bar_tick(dat, nTic)
}
\arguments{
\item{dat}{dat input with at least the following columns: Price, Size}

\item{nTic}{the number of ticks of each window}
}
\value{
time stamp at the end of each bar (if timestamp is provided), and H,L,O,C,V
}
\description{
Construct tick bars
}
\author{
Larry Lei Hua
}
