% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{bar_time}
\alias{bar_time}
\title{Construct time bars}
\usage{
bar_time(dat, tDur = 1)
}
\arguments{
\item{dat}{dat input with at least the following columns: tStamp, Price, Size}

\item{tDur}{the time duration in seconds of each window}
}
\description{
Construct time bars
}
