% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{bar_unit}
\alias{bar_unit}
\title{Construct unit bars}
\usage{
bar_unit(dat, unit)
}
\arguments{
\item{dat}{dat input with at least the following columns: Price, Size}

\item{unit}{the total dollar (unit) traded of each window}
}
\description{
Construct unit bars
}
