% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{bar_volume}
\alias{bar_volume}
\title{Construct volume bars}
\usage{
bar_volume(dat, vol)
}
\arguments{
\item{dat}{dat input with at least the following columns: Price, Size}

\item{vol}{the volume of each window}
}
\description{
Construct volume bars
}
