% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{bar_volume_imbalance}
\alias{bar_volume_imbalance}
\title{Construct volume imbalance bars}
\usage{
bar_volume_imbalance(dat, w0 = 10, bkw_T = 5, bkw_b = 5)
}
\arguments{
\item{dat}{dat input with at least the following column: Price, Size}

\item{w0}{the time window length of the first bar}

\item{bkw_T}{backward window length when using pracma::movavg for exponentially weighted average T}

\item{bkw_b}{backward window length when using pracma::movavg for exponentially weighted average b_tv_t}
}
\value{
a list of vectors for tStamp (if returned), and HLOCV of volume imbalance bars. Note that the remaining data after the latest imbalance time point will be formed as a bar.
}
\description{
Construct volume imbalance bars
}
\examples{

set.seed(1)
dat <- data.frame(Price = c(rep(0.5, 4), runif(50)), Size = rep(10,54))
bar_volume_imbalance(dat)

}
