% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.r
\name{card}
\alias{card}
\title{card}
\usage{
card(id = 0)
}
\arguments{
\item{id}{GPU id number.}
}
\value{
A card class object.
}
\description{
Constructor for GPU card objects.
}
\details{
Data is held in an external pointer.
}
