% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{condition_algoseek_equity_taq}
\alias{condition_algoseek_equity_taq}
\title{Obtain readable condition codes for AlgoSeek equity TAQ}
\usage{
condition_algoseek_equity_taq(con)
}
\arguments{
\item{con}{raw condition codes such as "a0000020" provided with Algoseek data}
}
\value{
a vector of integers for the condition codes defined by algoseek
}
\description{
Obtain readable condition codes for AlgoSeek equity TAQ
}
\examples{
condition_algoseek_equity_taq("a0000020")

}
\author{
Larry Lei Hua
}
