% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{condition_maps_algoseek_equity_taq}
\alias{condition_maps_algoseek_equity_taq}
\title{Map conditions to conditions codes for AlgoSeek equity TAQ}
\usage{
condition_maps_algoseek_equity_taq(codes, tq = "trade")
}
\arguments{
\item{codes}{a vector of integers for readable condition codes}

\item{tq}{a string to indicate "trade" or "quote"; default is "trade"}
}
\value{
a vector of meanings of the condition codes defined by algoseek
}
\description{
Map conditions to conditions codes for AlgoSeek equity TAQ
}
\examples{
condition_maps_algoseek_equity_taq(c(0, 9), tq="trade")
condition_maps_algoseek_equity_taq(c(0, 9), tq="quote")

}
\author{
Larry Lei Hua
}
