% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{cov}
\alias{cov}
\alias{stats_cov}
\alias{stats_cor}
\title{Covariance and Correlation}
\usage{
stats_cov(x, y = NULL, ret = NULL)

stats_cor(x, y = NULL, ret = NULL)
}
\arguments{
\item{x, y}{Input data.}

\item{ret}{The covariance/correlation matrix.}
}
\description{
Compute covariance and pearson correlation matrices.
}
