% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{cpu2cpu}
\alias{cpu2cpu}
\title{cpu2cpu}
\usage{
cpu2cpu(cpu_in, cpu_out)
}
\arguments{
\item{cpu_in}{Input.}

\item{cpu_out}{Ouput.}
}
\value{
Returns \code{NULL}.
}
\description{
Copy the data in one CPU object to another. The objects must be of the same
class (e.g. cpuvec/cpuvec and cpumat/cpumat), but they can be of different
fundamental types (e.g. float/double).
}
\examples{
library(fmlr)
x = cpumat(3, 2)
x$fill_linspace(1, 6)

y = cpumat(type="float")
cpu2cpu(x, y)

y$info()
y

}
