% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{cpu2gpu}
\alias{cpu2gpu}
\title{cpu2gpu}
\usage{
cpu2gpu(cpu_in, gpu_out)
}
\arguments{
\item{cpu_in}{Input.}

\item{gpu_out}{Ouput.}
}
\value{
Returns \code{NULL}.
}
\description{
Copy the data in a CPU object to a GPU object. The objects must be of the
core class (e.g. cpuvec/gpuvec and cpumat/gpumat), but they can be of
different fundamental types (e.g. float/double).
}
