% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{cpu2mpi}
\alias{cpu2mpi}
\title{cpu2mpi}
\usage{
cpu2mpi(cpu_in, mpi_out)
}
\arguments{
\item{mpi_in}{Input.}

\item{cpu_out}{Ouput.}

\item{rdest, cdest}{The row/column index of the receiving process.}
}
\value{
Returns \code{NULL}.
}
\description{
Copy the data in an MPI matrix to a CPU matrix. They can be of different
fundamental types (e.g. float/double).
}
