% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpumat-interface.r
\name{cpumat-class}
\alias{cpumat-class}
\alias{cpumatR6}
\title{cpumat class}
\description{
Storage and methods for CPU matrix data.
}
\details{
Data is held in an external pointer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cpumatR6$new()}}
\item \href{#method-resize}{\code{cpumatR6$resize()}}
\item \href{#method-set}{\code{cpumatR6$set()}}
\item \href{#method-info}{\code{cpumatR6$info()}}
\item \href{#method-print}{\code{cpumatR6$print()}}
\item \href{#method-fill_zero}{\code{cpumatR6$fill_zero()}}
\item \href{#method-fill_one}{\code{cpumatR6$fill_one()}}
\item \href{#method-fill_val}{\code{cpumatR6$fill_val()}}
\item \href{#method-fill_linspace}{\code{cpumatR6$fill_linspace()}}
\item \href{#method-fill_eye}{\code{cpumatR6$fill_eye()}}
\item \href{#method-fill_runif}{\code{cpumatR6$fill_runif()}}
\item \href{#method-fill_rnorm}{\code{cpumatR6$fill_rnorm()}}
\item \href{#method-scale}{\code{cpumatR6$scale()}}
\item \href{#method-rev_rows}{\code{cpumatR6$rev_rows()}}
\item \href{#method-rev_cols}{\code{cpumatR6$rev_cols()}}
\item \href{#method-nrows}{\code{cpumatR6$nrows()}}
\item \href{#method-ncols}{\code{cpumatR6$ncols()}}
\item \href{#method-dim}{\code{cpumatR6$dim()}}
\item \href{#method-data_ptr}{\code{cpumatR6$data_ptr()}}
\item \href{#method-to_robj}{\code{cpumatR6$to_robj()}}
\item \href{#method-from_robj}{\code{cpumatR6$from_robj()}}
\item \href{#method-clone}{\code{cpumatR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$new(nrows = 0, ncols = 0, type = "double")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nrows, ncols}}{The dimension of the matrix.}

\item{\code{type}}{Storage type for the matrix. Should be one of 'int', 'float', or 'double'.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Class initializer. See also \code{?cpumat}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resize"></a>}}
\subsection{Method \code{resize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$resize(nrows, ncols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nrows, ncols}}{The new dimension.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Change the dimension of the matrix object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$set(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{R matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the data in the cpumat object to point to the array in 'data'. See
also \code{?as_cpumat}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$info()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print one-line information about the matrix.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$print(ndigits = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ndigits}}{Number of decimal digits to print.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Print the data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_zero"></a>}}
\subsection{Method \code{fill_zero()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$fill_zero()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Fill all entries with zero.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_one"></a>}}
\subsection{Method \code{fill_one()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$fill_one()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Fill all entries with one.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_val"></a>}}
\subsection{Method \code{fill_val()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$fill_val(v)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v}}{Value to set all entries to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fill all entries with supplied value.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_linspace"></a>}}
\subsection{Method \code{fill_linspace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$fill_linspace(start, stop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start, stop}}{Beginning/end of the linear spacing.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fill the matrix (column-wise) with linearly-spaced values.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_eye"></a>}}
\subsection{Method \code{fill_eye()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$fill_eye()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Fill diagonal values to 1 and non-diagonal values to 0.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_runif"></a>}}
\subsection{Method \code{fill_runif()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$fill_runif(seed, min = 0, max = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{Seed for the generator. Can be left blank.}

\item{\code{min, max}}{Parameters for the generator.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fill the matrix with random unifmorm data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_rnorm"></a>}}
\subsection{Method \code{fill_rnorm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$fill_rnorm(seed, mean = 0, sd = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{Seed for the generator. Can be left blank.}

\item{\code{mean, sd}}{Parameters for the generator.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fill the matrix with random normal data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale"></a>}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$scale(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Value to scale all entries by.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Scale all entries by the supplied value.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rev_rows"></a>}}
\subsection{Method \code{rev_rows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$rev_rows()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Reverse rows.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rev_cols"></a>}}
\subsection{Method \code{rev_cols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$rev_cols()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Reverse columns.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nrows"></a>}}
\subsection{Method \code{nrows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$nrows()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns number of rows of the matrix.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ncols"></a>}}
\subsection{Method \code{ncols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$ncols()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns number of columns of the matrix.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\subsection{Method \code{dim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$dim()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns number of rows and columns of the matrix.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_ptr"></a>}}
\subsection{Method \code{data_ptr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$data_ptr()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the external pointer data. For developers only.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_robj"></a>}}
\subsection{Method \code{to_robj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$to_robj()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns an R matrix containing a copy of the class data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_robj"></a>}}
\subsection{Method \code{from_robj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$from_robj(robj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{robj}}{R matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Copies the values of the input to the class data. See also \code{?as_cpumat}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpumatR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
