% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpumat-interface.r
\name{cpumat}
\alias{cpumat}
\title{cpumat}
\usage{
cpumat(nrows = 0, ncols = 0, type = "double")
}
\arguments{
\item{nrows, ncols}{The dimensions of the matrix.}

\item{type}{Storage type for the matrix. Should be one of 'int', 'float', or 'double'.}
}
\value{
A cpumat class object.
}
\description{
Constructor for cpumat objects.
}
\details{
Data is held in an external pointer.
}
\seealso{
\code{\link{cpumat-class}}
}
