% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuvec.r
\name{cpuvec-class}
\alias{cpuvec-class}
\alias{cpuvecR6}
\title{cpuvec class}
\description{
Storage and methods for CPU vector data.
}
\details{
Data is held in an external pointer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cpuvecR6$new()}}
\item \href{#method-resize}{\code{cpuvecR6$resize()}}
\item \href{#method-inherit}{\code{cpuvecR6$inherit()}}
\item \href{#method-dupe}{\code{cpuvecR6$dupe()}}
\item \href{#method-info}{\code{cpuvecR6$info()}}
\item \href{#method-print}{\code{cpuvecR6$print()}}
\item \href{#method-fill_zero}{\code{cpuvecR6$fill_zero()}}
\item \href{#method-fill_val}{\code{cpuvecR6$fill_val()}}
\item \href{#method-fill_linspace}{\code{cpuvecR6$fill_linspace()}}
\item \href{#method-scale}{\code{cpuvecR6$scale()}}
\item \href{#method-rev}{\code{cpuvecR6$rev()}}
\item \href{#method-sum}{\code{cpuvecR6$sum()}}
\item \href{#method-get}{\code{cpuvecR6$get()}}
\item \href{#method-set}{\code{cpuvecR6$set()}}
\item \href{#method-size}{\code{cpuvecR6$size()}}
\item \href{#method-data_ptr}{\code{cpuvecR6$data_ptr()}}
\item \href{#method-get_type}{\code{cpuvecR6$get_type()}}
\item \href{#method-get_type_str}{\code{cpuvecR6$get_type_str()}}
\item \href{#method-to_robj}{\code{cpuvecR6$to_robj()}}
\item \href{#method-from_robj}{\code{cpuvecR6$from_robj()}}
\item \href{#method-clone}{\code{cpuvecR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$new(size = 0, type = "double")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The length of the vector.}

\item{\code{type}}{Storage type for the vector. Should be one of 'int', 'float', or 'double'.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Class initializer. See also \code{?cpuvec}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resize"></a>}}
\if{latex}{\out{\hypertarget{method-resize}{}}}
\subsection{Method \code{resize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$resize(size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{The new length.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Change the length of the vector object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inherit"></a>}}
\if{latex}{\out{\hypertarget{method-inherit}{}}}
\subsection{Method \code{inherit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$inherit(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{R vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the data in the cpuvec object to point to the array in 'data'. See
also \code{?as_cpuvec}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dupe"></a>}}
\if{latex}{\out{\hypertarget{method-dupe}{}}}
\subsection{Method \code{dupe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$dupe()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Duplicate the vector in a deep copy.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\if{latex}{\out{\hypertarget{method-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$info()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print one-line information about the vector.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$print(ndigits = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ndigits}}{Number of decimal digits to print.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Print the data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_zero"></a>}}
\if{latex}{\out{\hypertarget{method-fill_zero}{}}}
\subsection{Method \code{fill_zero()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$fill_zero()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Fill all entries with zero.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_val"></a>}}
\if{latex}{\out{\hypertarget{method-fill_val}{}}}
\subsection{Method \code{fill_val()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$fill_val(v)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{v}}{Value to set all entries to.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fill all entries with supplied value.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fill_linspace"></a>}}
\if{latex}{\out{\hypertarget{method-fill_linspace}{}}}
\subsection{Method \code{fill_linspace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$fill_linspace(start, stop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start, stop}}{Beginning/end of the linear spacing.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Fill the vector (column-wise) with linearly-spaced values.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale"></a>}}
\if{latex}{\out{\hypertarget{method-scale}{}}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$scale(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{Value to scale all entries by.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Scale all entries by the supplied value.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rev"></a>}}
\if{latex}{\out{\hypertarget{method-rev}{}}}
\subsection{Method \code{rev()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$rev()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Reverse rows.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sum"></a>}}
\if{latex}{\out{\hypertarget{method-sum}{}}}
\subsection{Method \code{sum()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$sum()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Sum the vector.
}

\subsection{Returns}{
Returns the sum.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$get(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Index (0-based).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get element from the vector.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$set(i, v)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{Index (0-based).}

\item{\code{v}}{Value.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set element of the vector.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$size()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns length of the vector.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-data_ptr}{}}}
\subsection{Method \code{data_ptr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$data_ptr()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the external pointer data. For developers only.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-get_type}{}}}
\subsection{Method \code{get_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$get_type()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the integer code for the underlying storage type. For developers only.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_type_str"></a>}}
\if{latex}{\out{\hypertarget{method-get_type_str}{}}}
\subsection{Method \code{get_type_str()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$get_type_str()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the string code for the underlying storage type. For developers only.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_robj"></a>}}
\if{latex}{\out{\hypertarget{method-to_robj}{}}}
\subsection{Method \code{to_robj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$to_robj()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns an R vector containing a copy of the class data.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_robj"></a>}}
\if{latex}{\out{\hypertarget{method-from_robj}{}}}
\subsection{Method \code{from_robj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$from_robj(robj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{robj}}{R vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Copies the values of the input to the class data. See also \code{?as_cpuvec}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cpuvecR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
