% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpuvec-interface.r
\name{cpuvec}
\alias{cpuvec}
\title{cpuvec}
\usage{
cpuvec(size = 0, type = "double")
}
\arguments{
\item{size}{Length of the vector.}

\item{type}{Storage type for the vector. Should be one of 'int', 'float', or 'double'.}
}
\value{
A cpuvec class object.
}
\description{
Constructor for cpuvec objects.
}
\details{
Data is held in an external pointer.
}
\seealso{
\code{\link{cpuvec-class}}
}
