% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ema}
\alias{ema}
\title{exponentially weighted moving average; only return the last value}
\usage{
ema(x, n)
}
\arguments{
\item{x}{a numeric vector}

\item{n}{window size}
}
\value{
a numeric value
}
\description{
exponentially weighted moving average; only return the last value
}
\author{
Larry Lei Hua
}
