% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracDiff.R
\name{fracDiff}
\alias{fracDiff}
\title{convert a time series into a fractionally differentiated series}
\usage{
fracDiff(x, d = 0.5, nWei = 10, tau = NULL)
}
\arguments{
\item{x}{a vector of time series to be fractionally differentiated}

\item{d}{the order for fractionally differentiated features}

\item{nWei}{number of weights for output}

\item{tau}{threshold where weights are cut off; default is NULL, if not NULL
then use tau and nWei is not used}
}
\description{
convert a time series into a fractionally differentiated series
}
\author{
Larry Lei Hua
}
