% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.r
\name{gpu-card-class}
\alias{gpu-card-class}
\alias{cardR6}
\title{card class}
\description{
Storage and methods for GPU card data.
}
\details{
Data is held in an external pointer.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cardR6$new()}}
\item \href{#method-set}{\code{cardR6$set()}}
\item \href{#method-info}{\code{cardR6$info()}}
\item \href{#method-print}{\code{cardR6$print()}}
\item \href{#method-get_id}{\code{cardR6$get_id()}}
\item \href{#method-valid_card}{\code{cardR6$valid_card()}}
\item \href{#method-data_ptr}{\code{cardR6$data_ptr()}}
\item \href{#method-clone}{\code{cardR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cardR6$new(id = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{GPU id number.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Class initializer.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cardR6$set(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{GPU id number.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set card to a different GPU.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\if{latex}{\out{\hypertarget{method-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cardR6$info()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print one-line information about the object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cardR6$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print one-line information about the object.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cardR6$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns GPU id number.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valid_card"></a>}}
\if{latex}{\out{\hypertarget{method-valid_card}{}}}
\subsection{Method \code{valid_card()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cardR6$valid_card()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns whether or not the card object is valid.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_ptr"></a>}}
\if{latex}{\out{\hypertarget{method-data_ptr}{}}}
\subsection{Method \code{data_ptr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cardR6$data_ptr()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the external pointer data. For developers only.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cardR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
