% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{gpu2gpu}
\alias{gpu2gpu}
\title{gpu2gpu}
\usage{
gpu2gpu(gpu_in, gpu_out)
}
\arguments{
\item{gpu_in}{Input.}

\item{gpu_out}{Ouput.}
}
\value{
Returns `NULL`.
}
\description{
Copy the data in one GPU object to another. The objects must be of the same
class (e.g. gpuvec/gpuvec and gpumat/gpumat), but they can be of different
fundamental types (e.g. float/double).
}
