% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpumat.r
\name{gpumat}
\alias{gpumat}
\title{gpumat}
\usage{
gpumat(card, nrows = 0, ncols = 0, type = "double")
}
\arguments{
\item{card}{A GPU card object; the return of \code{card()}. See \code{?card}.}

\item{nrows, ncols}{The dimensions of the matrix.}

\item{type}{Storage type for the matrix. Should be one of 'int', 'float', or 'double'.}
}
\value{
A gpumat class object.
}
\description{
Constructor for gpumat objects.
}
\details{
Data is held in an external pointer.
}
\seealso{
\code{\link{gpumat-class}}
}
