% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-interface.r
\name{grid}
\alias{grid}
\title{grid}
\usage{
grid(gridtype = PROC_GRID_SQUARE)
}
\arguments{
\item{id}{GPU id number.}
}
\value{
A grid class object.
}
\description{
Constructor for GPU grid objects.
}
\details{
Data is held in an external pointer.
}
