% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{imbalance_tick_v1}
\alias{imbalance_tick_v1}
\title{(Deprecated, use fmlr::imbalance_tick) The auxiliary function b_t for constructing tick imbalance bars. The first b_t is assigned the value 0 because no information is available}
\usage{
imbalance_tick_v1(dat)
}
\arguments{
\item{dat}{dat input with at least the following columns: Price}
}
\description{
(Deprecated, use fmlr::imbalance_tick) The auxiliary function b_t for constructing tick imbalance bars. The first b_t is assigned the value 0 because no information is available
}
\examples{

set.seed(1)
dat <- data.frame(Price = c(rep(0.5, 4), runif(10)))

b_t <- imbalance_tick_v1(dat)

}
