% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{istar_CUSUM}
\alias{istar_CUSUM}
\title{time index that triggers a symmetric CUSUM filter}
\usage{
istar_CUSUM(x, h)
}
\arguments{
\item{x}{a vector of time series to be filtered}

\item{h}{a vector of the thresholds}
}
\description{
time index that triggers a symmetric CUSUM filter
}
\examples{
set.seed(1)
x <- runif(100, 1, 3)
h <- rep(1.5, 100)
i_CUSUM <- istar_CUSUM(x,h)
abline(v=i_CUSUM, lty = 2)

## Comparing C and R versions
# set.seed(1)
# x <- runif(1000000, 1, 3)
# h <- rep(1.5, 100)

# start_time <- Sys.time()
# i_CUSUM <- istar_CUSUM(x,h)
# end_time <- Sys.time()
# C_time <- end_time - start_time

# start_time <- Sys.time()
# i_CUSUM_R <- istar_CUSUM_R(x,h)
# end_time <- Sys.time()
# R_time <- end_time - start_time
# cat("C and R time: ", C_time, R_time)
# all(i_CUSUM-i_CUSUM_R==0)
}
