% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_meta}
\alias{label_meta}
\title{Meta labeling, including three options: triple barriers, upper and vertical barriers, and lower and vertical barriers.}
\usage{
label_meta(x, events, ptSl, ex_vert = T, n_ex = 0)
}
\arguments{
\item{x}{a vector of prices series to be labeled.}

\item{events}{a dataframe that has the following columns:
\itemize{
         \item t0: event's starting time index.
         \item t1: event's ending time index; if t1==Inf then no vertical barrier, i.e., last observation in x is the vertical barrier.
         \item trgt: the unit absolute return used to set up the upper and lower barriers.
         \item side: 0: both upper and lower barriers; 1: only upper; -1: only lower.
        }}

\item{ptSl}{a vector of two multipliers for the upper and lower barriers, respectively.}

\item{ex_vert}{whether exclude the output when the vertical barrier is hit; default is T.}

\item{n_ex}{number of excluded observations at the begining of x; default is 0.}
}
\value{
data frame with the following columns:
\itemize{
        \item T_up: local time index when the upper barrier is hit; Inf means that upper is not hit.
        \item T_lo: local time index when the lower barrier is hit; Inf means that lower is not hit.
        \item t1: local time index when the vertical barrier is hit.
        \item ret: return associated with the event.
        \item label: low:-1, vertical:0, upper:1.
        \item t0Fea: begining time index of feature bars.
        \item t1Fea: ending time index of feature bars.
        \item tLabel: ending time index of events, i.e., when the labels are created. Both tFea and tLabel will be useful for sequential bootstrap.
        }
}
\description{
Meta labeling, including three options: triple barriers, upper and vertical barriers, and lower and vertical barriers.
}
