% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{add}
\alias{add}
\alias{linalg_add}
\title{add}
\usage{
linalg_add(
  transx = FALSE,
  transy = FALSE,
  x,
  y,
  ret = NULL,
  alpha = 1,
  beta = 1
)
}
\arguments{
\item{transx, transy}{Should x/y be transposed?}

\item{x, y}{Input data.}

\item{ret}{Either \code{NULL} or an already allocated fml matrix of the same
class and type as \code{x}.}

\item{alpha, beta}{Scalars.}
}
\value{
Returns the matrix sum.
}
\description{
Add two matrices: `ret = alpha*x + beta*y`.
}
