% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{chol}
\alias{chol}
\alias{linalg_chol}
\title{chol}
\usage{
linalg_chol(x)
}
\arguments{
\item{x}{Input data, overwritten by the inverse.}
}
\value{
Returns \code{NULL}.
}
\description{
Compute the lower-triangular Cholesky factor.
}
