% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{cpsvd}
\alias{cpsvd}
\alias{linalg_cpsvd}
\title{cpsvd}
\usage{
linalg_cpsvd(x, s, u = NULL, vt = NULL)
}
\arguments{
\item{x}{Input data. The input values are overwritten.}

\item{s}{Singular values.}

\item{u, vt}{The left/right singular vectors. Should both be \code{NULL} or
matrices of the same backend and fundamental type as \code{x}.}
}
\description{
"Crossproducts" SVD.
}
\details{
Computes the approximate SVD via the eigenvalue decomposition of
\code{crossprod(x)} if the input is tall/skinny and \code{tcrossprod(x)}
otherwise.
}
