% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{crossprod}
\alias{crossprod}
\alias{linalg_crossprod}
\alias{linalg_tcrossprod}
\title{crossprod}
\usage{
linalg_crossprod(alpha = 1, x, ret = NULL)

linalg_tcrossprod(alpha = 1, x, ret = NULL)
}
\arguments{
\item{alpha}{Number to scale the crossproduct by.}

\item{x}{Input data.}

\item{ret}{Either \code{NULL} or an already allocated fml matrix of the same
class and type as \code{x}.}
}
\value{
Returns the crossproduct.
}
\description{
Compute crossproducts.
}
