% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{det}
\alias{det}
\alias{linalg_det}
\title{det}
\usage{
linalg_det(x)
}
\arguments{
\item{x}{Input data, overwritten by its LU factorization.}
}
\value{
Returns a list containing the modulus and the sign.
}
\description{
Determinant
}
