% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{dot}
\alias{dot}
\alias{linalg_dot}
\title{dot}
\usage{
linalg_dot(x, y = NULL)
}
\arguments{
\item{x, y}{Input data.
class and type as \code{x}.}
}
\value{
Returns the dot product.
}
\description{
Compute vector dot product i.e. \code{sum(x*y)}.
}
