% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{eigen}
\alias{eigen}
\alias{linalg_eigen_sym}
\title{eigen}
\usage{
linalg_eigen_sym(x, values, vectors = NULL)
}
\arguments{
\item{x}{Input data. The input values are overwritten.}

\item{values}{The eigenvalues.}

\item{vectors}{The eigenvectors.}
}
\description{
Computes the eigenvalues and/or eigenvectors
}
\details{
You will need to initialize the return objects `values` and/or `vectors`
manually. See the example.
}
