% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{invert}
\alias{invert}
\alias{linalg_invert}
\title{invert}
\usage{
linalg_invert(x)
}
\arguments{
\item{x}{Input data, overwritten by the inverse.}
}
\value{
Returns `NULL`.
}
\description{
Invert a matrix.
}
