% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{lq_Q}
\alias{lq_Q}
\alias{linalg_lq_Q}
\title{lq_Q}
\usage{
linalg_lq_Q(LQ, lqaux, Q, work)
}
\arguments{
\item{LQ}{The compact LQ factorization. The return from \code{lq()}.}

\item{lqaux}{Auxilliary data for the compact LQ.}

\item{Q}{The output Q matrix.}

\item{work}{A workspace vector.}
}
\description{
Computes the Q matrix from the compact LQ factorization.
}
