% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{lq}
\alias{lq}
\alias{linalg_lq}
\title{lq}
\usage{
linalg_lq(x, lqaux)
}
\arguments{
\item{x}{Input data. The input values are overwritten.}

\item{lqaux}{Auxilliary data for the compact LQ.}
}
\description{
Computes the compact LQ factorization.
}
