% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{add}
\alias{add}
\alias{linalg_matmult}
\title{matmult}
\usage{
linalg_matmult(transx = FALSE, transy = FALSE, x, y, ret = NULL, alpha = 1)
}
\arguments{
\item{transx, transy}{Should x/y be transposed?}

\item{x, y}{Input data.}

\item{ret}{Either \code{NULL} or an already allocated fml matrix of the same
class and type as \code{x}.}

\item{alpha}{Scalar.}
}
\value{
Returns the matrix product.
}
\description{
Multiply two matrices: `ret = alpha*x*y`.
}
