% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{qr_Q}
\alias{qr_Q}
\alias{linalg_qr_Q}
\title{qr_Q}
\usage{
linalg_qr_Q(QR, qraux, Q, work)
}
\arguments{
\item{QR}{The compact QR factorization. The return from \code{qr()}.}

\item{qraux}{Auxilliary data for the compact QR.}

\item{Q}{The output Q matrix.}

\item{work}{A workspace vector.}
}
\description{
Computes the Q matrix from the compact QR factorization.
}
