% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{qr_R}
\alias{qr_R}
\alias{linalg_qr_R}
\title{qr_R}
\usage{
linalg_qr_R(QR, R)
}
\arguments{
\item{QR}{The compact QR factorization. The return from \code{qr()}.}

\item{R}{The output R matrix.}
}
\description{
Computes the R matrix from the compact QR factorization.
}
