% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{qr}
\alias{qr}
\alias{linalg_qr}
\title{qr}
\usage{
linalg_qr(x, qraux)
}
\arguments{
\item{x}{Input data. The input values are overwritten.}

\item{qraux}{Auxilliary data for the compact QR.}
}
\description{
Computes the compact QR factorization.
}
