% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{qrsvd}
\alias{qrsvd}
\alias{linalg_qrsvd}
\title{qrsvd}
\usage{
linalg_qrsvd(x, s, u = NULL, vt = NULL)
}
\arguments{
\item{x}{Input data. The input values are overwritten.}

\item{s}{Singular values.}

\item{u, vt}{The left/right singular vectors. Should both be \code{NULL} or
matrices of the same backend and fundamental type as \code{x}.}
}
\description{
QR/LQ-based SVD. Useful for very tall/skinny or short/wide data.
}
