% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{rsvd}
\alias{rsvd}
\alias{linalg_rsvd}
\title{rsvd}
\usage{
linalg_rsvd(seed, k, q, x, s, u = NULL, vt = NULL)
}
\arguments{
\item{seed}{Seed for the random generator.}

\item{k}{The number of components to estimate. Integer > 0.}

\item{q}{Exponent (see paper if you really care). Values of 1 or 2 are good.}

\item{x}{Input data. The input values are overwritten.}

\item{s}{Singular values.}

\item{u, vt}{The left/right singular vectors. Should both be \code{NULL} or
matrices of the same backend and fundamental type as \code{x}.}
}
\description{
SVD approximation via random projections.
}
\references{
Halko, Nathan, Per-Gunnar Martinsson, and Joel A. Tropp. "Finding
structure with randomness: Probabilistic algorithms for constructing
approximate matrix decompositions." SIAM review 53, no. 2 (2011): 217-288.
}
