% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{solve}
\alias{solve}
\alias{linalg_solve}
\title{solve}
\usage{
linalg_solve(x, y)
}
\arguments{
\item{x}{Input data. The input values are overwritten.}

\item{y}{The RHS, overwritten by the solution.}
}
\description{
Solve a system of equations.
}
