% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{svd}
\alias{svd}
\alias{linalg_svd}
\title{svd}
\usage{
linalg_svd(x, s, u = NULL, vt = NULL)
}
\arguments{
\item{x}{Input data. The input values are overwritten.}

\item{s}{Singular values.}

\item{u, vt}{The left/right singular vectors. Should both be \code{NULL} or
matrices of the same backend and fundamental type as \code{x}.}
}
\description{
Computes the singular value decomposition.
}
\details{
You will need to initialize the return objects \code{s} and/or \code{u} and
\code{vt}.
manually. See the example.
}
\examples{
suppressMessages(library(fmlr))
x = cpumat(3, 2)
x$fill_linspace(1, 6)

s = cpuvec()
linalg_svd(x, s)
s$info()
s$print()

}
