% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{trace}
\alias{trace}
\alias{linalg_trace}
\title{trace}
\usage{
linalg_trace(x)
}
\arguments{
\item{x}{Input data.}
}
\value{
Returns the trace.
}
\description{
Matrix trace, i.e. theh sum of the diagonal elements.
}
