% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{trinv}
\alias{trinv}
\alias{linalg_trinv}
\title{trinv}
\usage{
linalg_trinv(upper, unit_diag, x)
}
\arguments{
\item{upper}{Is the matrix upper triangular? Otherwise only the lower
triangle will be referenced.}

\item{unit_diag}{Is the matrix unit diagonal?}

\item{x}{Input data, overwritten by the inverse.}
}
\value{
Returns \code{NULL}.
}
\description{
Invert a triangular matrix.
}
