% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{xpose}
\alias{xpose}
\alias{linalg_xpose}
\title{xpose}
\usage{
linalg_xpose(x, ret = NULL)
}
\arguments{
\item{x}{Input data.}

\item{ret}{Either \code{NULL} or an already allocated fml matrix of the same
class and type as \code{x}.}
}
\value{
Returns the xpose.
}
\description{
Matrix transpose.
}
