% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{mpi2mpi}
\alias{mpi2mpi}
\title{mpi2mpi}
\usage{
mpi2mpi(mpi_in, mpi_out)
}
\arguments{
\item{mpi_in}{Input.}

\item{mpi_out}{Ouput.}
}
\value{
Returns \code{NULL}.
}
\description{
Copy the data in one MPI matrix to another. They can be of different
fundamental types (e.g. float/double).
}
