% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpimat.r
\name{mpimat}
\alias{mpimat}
\title{mpimat}
\usage{
mpimat(grid, nrows = 0, ncols = 0, bf_rows = 16, bf_cols = 16, type = "double")
}
\arguments{
\item{grid}{An MPI grid object; the return of \code{grid()}. See \code{?grid}.}

\item{nrows, ncols}{The dimensions of the matrix.}

\item{bf_rows, bf_cols}{The blocking factor.}

\item{type}{Storage type for the matrix. Should be one of 'int', 'float', or 'double'.}
}
\value{
A mpimat class object.
}
\description{
Constructor for mpimat objects.
}
\details{
Data is held in an external pointer.
}
\seealso{
\code{\link{mpimat-class}}
}
