% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{pca}
\alias{pca}
\alias{stats_pca}
\title{pca}
\usage{
stats_pca(rm_mean = TRUE, rm_sd = FALSE, x, sdev, rot = NULL)
}
\arguments{
\item{rm_mean, rm_sd}{Should the data be centered/scaled first? Operations
occur in-place.}

\item{x}{Input matrix. The input values are overwritten.}

\item{sdev}{The standard deviations for the principal components.}

\item{rot}{The pca loadings. If \code{NULL} is passed, then these will not
be calculated.}
}
\description{
Perform principal components analysis.
}
\examples{
suppressMessages(library(fmlr))
x = cpumat(3, 2)
x$fill_linspace(1, 6)

s = cpuvec()
stats_pca(TRUE, TRUE, x, s)
s$info()
s$print()

}
