% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purgedCV.R
\name{purged_k_CV}
\alias{purged_k_CV}
\title{Purged k-fold CV with embargo}
\usage{
purged_k_CV(feaMat, k = 5, gam = 0.01)
}
\arguments{
\item{feaMat}{a data.frame for feature matrix with the first column being the label}

\item{k}{number of folds for k-fold CV}

\item{gam}{gamma for embargo}
}
\value{
a list of k data.frame, each containing a test set and a training set
}
\description{
Purged k-fold CV with embargo
}
\examples{
feaMat <- data.frame(Y = c(1,1,0,1,0),
                     V = c(2,4,2,4,1), 
                     t1Fea = c(2,5,8,14,20), 
                     tLabel = c(4,12,16,23,38))
purged_k_CV(feaMat, k=2, gam=0.1)

}
\author{
Larry Lei Hua
}
