% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{read_algoseek_futures_fullDepth}
\alias{read_algoseek_futures_fullDepth}
\title{Load AlgoSeek Futures Full Depth data from zip files}
\usage{
read_algoseek_futures_fullDepth(zipdata, whichData = NULL)
}
\arguments{
\item{zipdata}{the original zip data provided by AlgoSeek}

\item{whichData}{the specific data to be loaded; by default load all data in the zip file}
}
\description{
Load AlgoSeek Futures Full Depth data from zip files
}
\examples{
\donttest{
zipdata <- tempfile()
download.file("https://www.algoseek.com/static/files/sample_data/
futures_and_future_options/ESH5.Futures.FullDepth.20150128.csv.zip",zipdata)
dat <- read_algoseek_futures_fullDepth(zipdata)
}

# Do not run unless the file 20160104.zip is avaliable
# dat <- read_algoseek_futures_fullDepth("20160104.zip", whichData="ES/ESH6.csv")

}
\author{
Larry Lei Hua
}
