% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimops.r
\name{scale}
\alias{scale}
\alias{dimops_scale}
\title{scale}
\usage{
dimops_scale(rm_mean = TRUE, rm_sd = FALSE, x)
}
\arguments{
\item{rm_mean, rm_sd}{Should the data be centered/scaled first?}

\item{x}{Input matrix. The input data is overwritten by the centered/scaled
data.}
}
\description{
Remove the rows and/or sd of the columns of a matrix. The operations occur
in-place.
}
