% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{Tstar_trb}
\alias{Tstar_trb}
\title{Tstar index for Tick Runs Bars (bar_trb)}
\usage{
Tstar_trb(dat, w0 = 10, bkw_T = 5, bkw_Pb1 = 5)
}
\arguments{
\item{dat}{dat input with at least the following columns: Price}

\item{w0}{the time window length of the first bar}

\item{bkw_T}{backward window length when using pracma::movavg for exponentially weighted average T}

\item{bkw_b}{backward window length when using pracma::movavg for exponentially weighted average b_t}
}
\value{
a vector for the lengths of the tick imbalance bars. For example, if the return is c(10,26), then the 2 tick imbalance bars are (0,10] and (10, 36]
}
\description{
Tstar index for Tick Runs Bars (bar_trb)
}
\examples{

set.seed(1)
dat <- data.frame(Price = c(rep(0.5, 4), runif(50)))
T_trb <- Tstar_trb(dat)
b_t <- imbalance_tick(dat)
cumsum(b_t)[cumsum(T_trb)] # check the accumulated b_t's where the imbalances occur

}
