% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Tstar_trb_cpp}
\alias{Tstar_trb_cpp}
\title{Tstar index for Tick Runs Bars (bar_trb)}
\usage{
Tstar_trb_cpp(b_t, w0, de, bkw_T, bkw_Pb1)
}
\arguments{
\item{b_t}{output of imbalance_tick(dat) with the dat has at least the following columns: Price}

\item{w0}{the time window length of the first bar}

\item{de}{a positive value for adjusting the expected window size, that is, de*E0}

\item{bkw_T}{backward window length for exponentially weighted average T}

\item{bkw_Pb1}{backward window length for exponentially weighted average P[b_t=1]}
}
\value{
a list of the following two vectors: 
    a vector for the lengths of the tick imbalance bars. For example, if the return is c(10,26), then the 2 tick imbalance bars are (0,10] and (10, 36]
    a vector indicating up runs or down runs
}
\description{
Tstar index for Tick Runs Bars (bar_trb)
}
\examples{

set.seed(1)
dat <- data.frame(Price = c(rep(0.5, 5), runif(100)))
b_t <- imbalance_tick(dat)
T_trb <- Tstar_trb_cpp(b_t, 10, 1.0, 10, 10)
col <- ifelse(T_trb$Type==1, "red", "blue")
T <- cumsum(T_trb$Tstar)
plot(dat$Price)
for(i in 1:length(T)) abline(v = T[i], col = col[i])
 
}
