% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{Tstar_vib}
\alias{Tstar_vib}
\title{Tstar index for Volume Imbalance Bars (bar_vib)}
\usage{
Tstar_vib(dat, w0 = 10, bkw_T = 5, bkw_b = 5)
}
\arguments{
\item{dat}{dat input with at least the following columns: Price, Size}

\item{w0}{the expected time window length of the first bar}

\item{bkw_T}{backward window length when using exponentially weighted average T}

\item{bkw_b}{backward window length when using exponentially weighted average b_tv_t}
}
\value{
a vector for the lengths of the tick imbalance bars. For example, if the return is c(10,26), then the 2 tick imbalance bars are (0,10] and (10, 36]
}
\description{
Tstar index for Volume Imbalance Bars (bar_vib)
}
\examples{

set.seed(1)
dat <- data.frame(Price = c(rep(0.5, 4), runif(50)), Size = rep(10,54))
T_vib <- Tstar_vib(dat)
b_tv_t <- imbalance_volume(dat)
cumsum(b_tv_t)[cumsum(T_vib)] # check the accumulated b_tv_t's where the imbalances occur

}
\author{
Larry Lei Hua
}
