% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{Tstar_vrb}
\alias{Tstar_vrb}
\title{Tstar index for Volume Runs Bars (bar_vrb)}
\usage{
Tstar_vrb(dat, w0 = 10, bkw_T = 5, bkw_Pb1 = 5)
}
\arguments{
\item{dat}{dat input with at least the following columns: Price, Size}

\item{w0}{the time window length of the first bar}

\item{bkw_T}{backward window length when using pracma::movavg for exponentially weighted average T}

\item{bkw_Pb1}{backward window length when using pracma::movavg for exponentially weighted average P[b_t=1]}
}
\value{
a vector for the lengths of the tick runs bars. For example, if the return is c(10,26), then the 2 tick runs bars are (0,10] and (10, 36]
}
\description{
Tstar index for Volume Runs Bars (bar_vrb)
}
\examples{

set.seed(1)
dat <- data.frame(Price = c(rep(0.5, 4), runif(50)),
                  Size = rep(10,54))
T_vrb <- Tstar_vrb(dat)

}
