% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracDiff.R
\name{weights_fracDiff}
\alias{weights_fracDiff}
\title{calculate the weights for deriving fractionally differentiated series}
\usage{
weights_fracDiff(d = 0.5, nWei = 10, tau = NULL)
}
\arguments{
\item{d}{the order for fractionally differentiated features}

\item{nWei}{number of weights for output}

\item{tau}{threshold where weights are cut off; default is NULL, if not NULL then use tau and nWei is not used}
}
\description{
calculate the weights for deriving fractionally differentiated series
}
\examples{
weights_fracDiff(0.5,tau=1e-3)

}
