% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{acc_lucky}
\alias{acc_lucky}
\title{A function to check whether a classification is better than guess}
\usage{
acc_lucky(train_class, test_class, my_acc, s = 1000)
}
\description{
A function to check whether a classification is better than guess
}
