% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimops.r
\name{dimops}
\alias{dimops}
\alias{dimops_rowsums}
\alias{dimops_rowmeans}
\alias{dimops_colsums}
\alias{dimops_colmeans}
\title{matsums}
\usage{
dimops_rowsums(x, s = NULL)

dimops_rowmeans(x, s = NULL)

dimops_colsums(x, s = NULL)

dimops_colmeans(x, s = NULL)
}
\arguments{
\item{x}{Input matrix.}

\item{s}{Either \code{NULL} or an already allocated fml matrix of the same
class and type as \code{x}.}
}
\value{
Returns the matrix sum.
}
\description{
Compute the sum of rows.
}
