% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmlr-package.r
\docType{package}
\name{fmlr-package}
\alias{fmlr-package}
\title{fmlr}
\description{
fmlr is an R interface to the fml library. It is a
"medium-level" interface for multiple dense matrix types, principally CPU,
GPU, and MPI. Each supports multiple fundamental types (int, float, double),
and data is held externally to R and operations that modify data generally
occur in-place. The interface largely tracks with the core fml interface.
We use R6 so that generally an R code can be easily translated to C++.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
