% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuvec-interface.r
\name{gpuvec}
\alias{gpuvec}
\title{gpuvec}
\usage{
gpuvec(card, size = 0, type = "double")
}
\arguments{
\item{size}{Length of the vector.}

\item{type}{Storage type for the vector. Should be one of 'int', 'float', or 'double'.}
}
\value{
A gpuvec class object.
}
\description{
Constructor for gpuvec objects.
}
\details{
Data is held in an external pointer.
}
\seealso{
\code{\link{gpuvec-class}}
}
