% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.r
\name{grid}
\alias{grid}
\title{grid}
\usage{
grid(gridtype = PROC_GRID_SQUARE)
}
\arguments{
\item{gridtype}{Type of processor grid: \code{PROC_GRID_SQUARE},
\code{PROC_GRID_WIDE}, or \code{PROC_GRID_TALL}.}
}
\value{
A grid class object.
}
\description{
Constructor for MPI grid objects.
}
\details{
Data is held in an external pointer.
}
