% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{istar_CUSUM_R}
\alias{istar_CUSUM_R}
\title{time index that triggers a symmetric CUSUM filter (R version for istar_CUSUM())}
\usage{
istar_CUSUM_R(x, h)
}
\arguments{
\item{x}{a vector of time series to be filtered}

\item{h}{a positive number for the thresholds}
}
\description{
time index that triggers a symmetric CUSUM filter (R version for istar_CUSUM())
}
