% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{cond}
\alias{cond}
\alias{linalg_cond_1}
\title{Condition Number}
\usage{
linalg_cond_1(x)
}
\arguments{
\item{x}{Input data. The data is modified in each case.}
}
\value{
The requested condition number.
}
\description{
Condition numbers.
}
