% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.r
\name{trace}
\alias{trace}
\alias{linalg_svd}
\title{svd}
\usage{
linalg_svd(x, s, u = NULL, vt = NULL)
}
\arguments{
\item{x}{Input data. The input values are overwritten.}

\item{s}{Singular values.}

\item{u, vt}{The left/right singular vectors. Should both be `NULL` or
matrices of the same backend and fundamental type as `x`.}
}
\description{
Computes the singular value decomposition.
}
\details{
You will need to initialize the return objects `s` and/or `u` and `vt`
manually. See the example.
}
\examples{
suppressMessages(library(fmlr))
x = cpumat(3, 2)
x$fill_linspace(1, 6)

s = cpuvec()
linalg_svd(x, s)
s$info()
s$print()

}
