% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimops.r
\name{scale}
\alias{scale}
\alias{dimops_scale}
\title{scale}
\usage{
dimops_scale(rm_mean = TRUE, rm_sd = FALSE, x)
}
\arguments{
\item{x}{Input matrix.}
}
\value{
Returns the matrix sum.
}
\description{
Remove the rows and/or sd of the columns of a matrix.
}
